refactor(Role): Player만 베팅 금액을 가지도록 변경
