+++ src/main/kotlin/blackjack/model/playing/participants/Dealer.kt
@@ -2,9 +2,8 @@ package blackjack.model.playing.participants
 
 import blackjack.model.playing.cardhand.CardHand
 import blackjack.model.playing.participants.player.PlayerName
-import blackjack.model.winning.Betting
 
-data class Dealer(override val cardHand: CardHand) : Role(name = PlayerName(DEALER), cardHand, betting = Betting(0)) {
+data class Dealer(override val cardHand: CardHand) : Role(name = PlayerName(DEALER), cardHand) {
     override fun canDraw(): Boolean = cardHand.calculateScore() < DEALER_MAX_HIT_SUM
 
     companion object {
