+++ src/main/kotlin/blackjack/model/playing/participants/Role.kt
@@ -4,10 +4,9 @@ import blackjack.model.card.Card
 import blackjack.model.card.CardDeck
 import blackjack.model.playing.cardhand.CardHand
 import blackjack.model.playing.participants.player.PlayerName
-import blackjack.model.winning.Betting
 import blackjack.model.winning.WinningResultStatus
 
-abstract class Role(open val name: PlayerName, open val cardHand: CardHand, open val betting: Betting) {
+abstract class Role(open val name: PlayerName, open val cardHand: CardHand) {
     abstract fun canDraw(): Boolean
 
     fun addInitialCards(cardDeck: CardDeck) {
@@ -33,25 +32,8 @@ abstract class Role(open val name: PlayerName, open val cardHand: CardHand, open
         }
     }
 
-    fun calculateProfit(winningResultStatus: WinningResultStatus): Double {
-        return when (winningResultStatus) {
-            WinningResultStatus.PUSH -> PUSH_MULTIPLIER
-            WinningResultStatus.DEFEAT -> betting.amount * DEFEAT_MULTIPLIER
-            WinningResultStatus.VICTORY ->
-                if (cardHand.isBlackjack(this)) {
-                    betting.amount * BLACK_JACK_MULTIPLIER
-                } else {
-                    betting.amount * VICTORY_MULTIPLIER
-                }
-        }
-    }
-
     companion object {
         private const val INITIAL_CARDS_COUNT = 2
         private const val BLACK_JACK_SCORE = 21
-        private const val BLACK_JACK_MULTIPLIER = 1.5
-        private const val VICTORY_MULTIPLIER = 1.0
-        private const val DEFEAT_MULTIPLIER = -1.0
-        private const val PUSH_MULTIPLIER = 0.0
     }
 }
